Destroy asteroid when it gets too low and hits disposal station
When asteroid destruction commences, the asteroid model is swapped for one that looks identical but is comprised of pieces. The pieces break apart using Unity's physics. For each piece, a particle system with sprites is instantiated, while the pieces themselves are shrunk over time, then their game objects destroyed as they disappear. The effect is of the asteroids breaking into tinier and tinier pieces until it is space dust.